Removed a Hide_TensorFlow_Content element as it was empty
training.html file has 4 elements namely: hide pytorch content (2) and hide tensorflow content(2).
Both of the "hide pytorch content" elements work fine. One of the "hide tensorflow content" element
has no content. Hence when one of those "hide tensorflow content" are clicked, the second element
shows up empty.

This solution removes the second "hide tensorflow content" element as it has no information to
display. There are not really an side effects or unintuitive consequences of this change apart
from the fact that there will be two pytorch elements and one tensorflow element, which are just
design related issues.

A side note: the original author of the documentation might have forgotten to add information
to the second "hide tensorflow content" element. But it is very unlikely because tensorflow
and pytorch are different when it comes to custom training loops (which is what its about).